Machine learning algorithm and one application implementation.
It contains K Means and K-NN algorithm code and shows how it works and
how we can use it in our code using the classes provided.

And the machine learning application which is based on the analysis of
different rounds played in this program, strategy one ( FOne class)
works the best for me in terms of profit.